+++ packages/address/src/index.ts
@@ -1,4 +1,4 @@
-export * from './bb/MockAddress';
+export * from './mock/MockAddress';
 export * from './Address';
 export * from './ImmutableAddress';
 export * from './MutableAddress';
