+++ packages/project/src/index.ts
@@ -1,4 +1,4 @@
-export * from './bb/MockProject';
+export * from './mock/MockProject';
 export * from './ImmutableProject';
 export * from './MutableProject';
 export * from './Project';
