+++ packages/sequence/src/index.ts
@@ -1,4 +1,4 @@
-export * from './bb/MockSequence';
+export * from './mock/MockSequence';
 export * from './ImmutableSequence';
 export * from './MutableSequence';
 export * from './ReadonlySequence';
