+++ packages/sequence/src/mock/MockSequence.ts
