+++ packages/tree/src/ClosureTable/index.ts
@@ -1,4 +1,4 @@
-export * from './bb/index';
+export * from './mock/index';
 export * from './ClosureTable';
 export * from './ClosureTableHierarchies';
 export * from './ClosureTableHierarchy';
