+++ packages/tree/src/ClosureTable/mock/MockTreeIDFactory.ts
@@ -1,4 +1,4 @@
-import { MockTreeID } from '../../bb/MockTreeID';
+import { MockTreeID } from '../../mock/MockTreeID';
 import { TreeIDFactory } from '../TreeIDFactory';
 
 export class MockTreeIDFactory implements TreeIDFactory<MockTreeID> {
