+++ packages/tree/src/TreeNode/__tests__/SerializableTreeNode.spec.ts
@@ -1,6 +1,6 @@
 import { ImmutableAddress, MockAddress, MutableAddress } from '@jamashita/lluvia-address';
-import { MockTreeID } from '../../bb/MockTreeID';
-import { MockTreeObject } from '../../bb/MockTreeObject';
+import { MockTreeID } from '../../mock/MockTreeID';
+import { MockTreeObject } from '../../mock/MockTreeObject';
 import { SerializableTreeNode } from '../SerializableTreeNode';
 
 describe('SerializableTreeNode', () => {
